Split the report screen into different views.
This adds many more files to the application, but there's a reduction in
complexity of each component, and all of these components interact in
ways described by ember.js' docs, so I think this is a win.

This commit adds in fixtures for testing purposes; these fixtures are
always included in the asset pipeline.  It's not necessarily a huge
problem if they're in production builds, but it'd be nice if there was a
way to only include them in development situations.

Finally, the event search criteria screen is still something of a mess,
but that can be tackled later.