Switch to history-based routing.
This makes it easier to bookmark URLs and come back to them at a later
time, which goes along very well with the search-as-resource paradigm
introduced in previous commits.

Browser compatibility isn't as good as the anchor-based method, but
we're really only supporting recent-ish versions of Firefox for now, and
I know that this works in Chrome and Safari too.  (What that really
means: lower versions of IE are out.)