More resource rearrangement.
I read Ember.js' router code over the weekend, and it occurred to me
that the routing guides on emberjs.com didn't really tell the whole
story.  In particular I did not previously know that this was possible:

  @resource 'event_searches', ->
    @resource 'show', path: '/:event_search_id'

Now, all of that is *inferrable* from the guides, but not spelled out.
(What can I say, I'm not very good at inference.)

Anyway, after reading said code, I was inspired to clean up the event
search routes.

This moves things much more in line towards one-controller-per-outlet
(not a requirement, but often a useful pattern).