Introduce Pancakes.Tagger.
Although I have indeed had good experiences with tag-it, there are a
couple of things it does that made it evident (to me) that it was not
appropriate for this use:

1. tag-it assumes that it has full control of the DOM subtree on which
   it takes effect.  This makes populating said tree with data from
   Ember models very messy.
2. tag-it operates only on strings, not arbitrary Javascript objects.
   However, in this app, it is much easier to express tagging as
   addition and removal of objects to a collection.

After spending a few hours looking for a clean way to hook up tag-it to
manipulate the contents of the eventTypes association on EventSearch, I
gave up and just started writing something in (my best understanding of)
The Ember.js Way.

The result's actually pretty short.  (Which I guess one would expect
after dragging in a framework like Ember.js.)

I still need to apply component styles, and Pancakes.Tagger itself could
probably use some polish.  (What doesn't?)  But it works for everything
I need it to do.

Oh yeah, tests too.  Hmm.