More formal reload/caching behavior for data stores.
Specifically, store contents are now asynchronously loaded at
application startup.  They may be reloaded at any point in time.
When the store contents are requested, the store waits for all load
requests to complete and then returns the cached result.

The benefit of this approach is that it makes expected behavior of the
PUT /api/v1/mdes_version call implementable without dealing with a bunch
of ugly code to check cache state.  Instead, we leave that to Celluloid's
futures.

This also lets us load rarely-changing data in parallel with application
startup, which can be a win for startup time and responsiveness.