Love EventType and StudyLocation for what they are.
Ember Data expects the ID to map to "id".  It doesn't have to, though,
which is quite convenient, as neither EventType nor StudyLocation have
an "id" property.  (They are uniquely identifiable, but "id" is not the
appropriate property name in either case.)

Also camelCased EventType's properties to give it more of that DS.Model
feel.