Move towards client-side UUID generation for EventSearches.
The main reason is that extracting an ID from a server payload is more
fiddly work than just stating "oh, we have nothing to read back".

This commit is (a lot) more code, but I think it's worth it.  It means
that the client now truly needs _only_ an acknowledgment from the server
to know that its view of the world is in sync with the server's; no ID
munging is required.

This commit also unifies the create and update actions on
EventSearchesController, since they're now the same action.