Send status URL as a Link header.
There's a long story behind this one, so I'm going to just write it out
here.

Ember Data's current support for embedded has-many associations is
really spotty.  It will serialize them just fine, but reading them back
into a just-serialized record just doesn't seem to work at all right
now.  The short version is that associated records do not seem to fully
participate in the same state transitions as their parent, which means
that an associated record will be in state inFlight (or uncommitted) and
suddenly receive a loadedData event when server-supplied data arrives.
That event isn't legal in either of those states, and we get an
exception.

As far as I can tell, that's an Ember Data bug.  But it's something that
I've got to live with.  (Dumping Ember Data is a possibility, but I
think it's too late for that.  Lesson learned for next iteration.)

HOWEVER, if we return *nothing*, then Ember Data's RESTAdapter doesn't
modify the record's contents at all, which sidesteps the symptom.  (Are
the associated records in the wrong state still?  Yeah, probably.  I'll
find out soon enough if that'll screw anything else up.)

Hence this change.  We now return nothing (which is the default behavior
of the ActionController responder on a PUT, anyway) and throw the
associated URLs in a Link header.  This is, to the best of my knowledge,
an acceptable use of that header.

(One other nice side effect: status URLs will no longer go stale if e.g.
the app hostname changes.)

This prevents the state transition errors in Ember Data, but -- as I
said -- likely doesn't fix them.  It does buy me some time to file a
proper bug report and/or figure out how I'm holding it wrong, though.

The next bit of this data-to-header conversion is to extend
Pancakes.Adapter and Pancakes.EventSearch to respectively fetch and
store the status URL.