Store last start time in an EventSearch.
A single event search may be run multiple times in parallel (i.e. if
it's changed and resubmitted while another job is running).

The GET /api/v1/event_searches/:id/status endpoint must always return
the status of the latest started job.  To identify that job, we need the
job ID (which is the search ID) and the time when it was started.