Do not queue event search jobs on PUT.
A user may perform a GET on any event search at any time; however, it is
possible that we will not be able to service the request out of cache,
due to data expiration.  We therefore need an endpoint to explicitly
trigger a refresh.  However, once we've got that, there's not really any
reason left to queue items from two endpoints.

Well, "avoiding round-trips" could be a reason.  I'll have to see how
badly that one extra request per search hurts things.