Update query set status.
Yes, there's a lot more going on here than is necessary to satisfy the
"when a search is started, progress is updated" feature.  Even better,
none of this new code has tests.

The justifications exist, but they're weak, so I won't bother with
them.  Instead, I'll just describe how this code works.

Status updates consist of two parts.  The first part addresses the
status of the query set as a whole; the second part addresses elements
in the query set.  The first part is set on the search object and is the
basis for the values of the isStarted and isDone properties.  The second
part is also set on the search object, but is also sent to a
LocationStatusController, which is used to render a status glyph for
each location.

Other data flows:

1. When we enter the event_searches.show route (either from .new or
   direct entry), we end up with an EventSearch model.  This model
   becomes EventSearchesShowController's content.

2. When the search finishes loading, we request its status.  Once
   we've gotten that, we check whether or not the server has started the
   search job.  If it hasn't, we kick it off.

3. Kicking off a status update also updates an isDone property on the
   search model.  If that property is false, we kick off a poller timer to
   periodically update status.

To do:

In flow #2, "started" isn't quite the right criterion.  We really only
want to kick off a job if (a) a refresh was explicitly requested or (b)
the job hasn't been *queued* yet.  Implementing (b) will require a bit
more server work.