Juggle around authority setup.
Reason: When authorities are present in an Aker.configure block (which
is what config.aker eventually passes its block to), the authority
initialization process starts.  Part of the initialization process for
the CAS authority involves checking whether required parameters are
present, which means that we can't install required authorities until we
know where our parameters are coming from.