Lock out search button when no date range is present.  Fixes #4160.
There's a compromise in this commit which is due to (what I think is) a
limitation of Hamlbars' binding helpers.

Specifically, the bindings for bound attributes execute in a different
context than bindings to properties, and I can't find a way to reconcile
the two contexts.  (That's why the disabled/enabled properties are on
the model, not the controller.)