Make jQuery UI's datepicker enter dates in YYYY-MM-DD.  #4172.
We expect YYYY-MM-DD format, so might as well make the datepicker do
that, too.

It'd be nice to have a single place for date formatting.  It's possible
to do something like this:

    $.datepicker
      onSelect: (date, dp) =>
        @set 'value', moment(date)

but while this does help us maintain a consistent format, it introduces
undesirable visual glitches as the datepicker's value is updated twice
in quick succession.