Timecode reports by EventSearch#updated_at, not #last_started_at.
There is a time lag between a job submission and Sidekiq starting a job.
This time lag can be long enough that Pancakes' frontend makes a status
query and finds the old status of the report.

If this occurs with previously-completed reports, Pancakes' frontend
will not go into "waiting for results" mode.